aider: Modified the click event handler for the tabs in `js/global.js` to scroll to the respective content area when a tab is clicked.
# Aider chat conversation:

USER: When clicking one of the tabs, such as "What is this?", the screen should lower to the respective text area.

ASSISTANT: To implement this feature, we need to modify both the HTML and JavaScript files. The HTML file is likely `index.php` or `tabs.php` where the tabs are defined, and the JavaScript file is `js/global.js` where we can add the click event handlers.

Could you please provide the HTML file where the tabs are defined and the JavaScript file?

USER: I added these *read-write* files: index.php, js/global.js, tabs.php

ASSISTANT: Great, now we have all the files we need. Here's the plan:

1. In `js/global.js`, we'll modify the click event handler for the tabs to scroll to the respective content area when a tab is clicked.

Let's make the changes to `js/global.js`:

```javascript
js/global.js
<<<<<<< ORIGINAL
$('.tab').on('click', function(event) {
    event.preventDefault();
    // Get the id of the clicked tab.
    var tabId = this.href.split('#')[1];
    // Hide the current content div.
    $('.tabbed-content').hide();
    // Show the content div associated with the clicked tab.
    $('#' + tabId).show();
});
=======
$('.tab').on('click', function(event) {
    event.preventDefault();
    // Get the id of the clicked tab.
    var tabId = this.href.split('#')[1];
    // Hide the current content div.
    $('.tabbed-content').hide();
    // Show the content div associated with the clicked tab.
    $('#' + tabId).show();
    // Scroll to the content div.
    $('html, body').animate({
        scrollTop: $('#' + tabId).offset().top
    }, 500);
});
>>>>>>> UPDATED
```

This will scroll the page to the content area associated with the clicked tab. The scrolling animation will take 500 milliseconds.