aider: Modified click event handler in `js/global.js` to only prevent default event if the href starts with '#'.
# Aider chat conversation:

USER: In safari, the tab clicks aren't moving the screen. It only works in Chrome.

USER: In safari, the tab clicks aren't moving the screen. It only works in Chrome.

ASSISTANT: This issue might be due to differences in how Safari and Chrome handle JavaScript's `event.preventDefault()`. In Safari, it might be preventing the default scroll behavior.

To fix this, we need to modify the JavaScript file `js/global.js` where the click event handlers for the tabs are defined.

Could you please add `js/global.js` to the chat so we can make the necessary changes?

USER: I added these *read-write* files: js/global.js

ASSISTANT: Great, now we have the file we need. Here's the plan:

1. In `js/global.js`, we'll modify the click event handler for the tabs to only prevent the default event if the href of the clicked tab starts with '#'. This should allow the default scroll behavior in Safari while still preventing the page from reloading.

Let's make the changes to `js/global.js`:

```javascript
js/global.js
<<<<<<< ORIGINAL
$('.tab').on('click', function(event) {
    event.preventDefault();
    // Get the id of the clicked tab.
    var tabId = this.href.split('#')[1];
    // Hide the current content div.
    $('.tabbed-content').hide();
    // Show the content div associated with the clicked tab.
    $('#' + tabId).show();
    // Scroll to the content div.
    $('html, body').animate({
        scrollTop: $('#' + tabId).offset().top
    }, 500);
});
=======
$('.tab').on('click', function(event) {
    // Only prevent the default event if the href starts with '#'
    if (this.href.startsWith('#')) {
        event.preventDefault();
    }
    // Get the id of the clicked tab.
    var tabId = this.href.split('#')[1];
    // Hide the current content div.
    $('.tabbed-content').hide();
    // Show the content div associated with the clicked tab.
    $('#' + tabId).show();
    // Scroll to the content div.
    $('html, body').animate({
        scrollTop: $('#' + tabId).offset().top
    }, 500);
});
>>>>>>> UPDATED
```

This should fix the issue in Safari while still working in Chrome. Please test it and let me know if it works.